requirements.txt
pika>=1.1
python-dotenv==0.21.0